Creacion de roles para la aplicacion
-se hicieron roles administrador, y cliente.
-se creo la pantalla para el cliente, y modifico la parte del admin para que funicone el boton
