+++ next.config.js
@@ -1,5 +1,6 @@
 const withPWA = require('next-pwa')({
     dest: 'public',
+    disable: process.env.NODE_ENV === "development",
     register: true,
     skipWaiting: true,
   })
