+++ src/app/clientes/alquiler/page.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function alquiler() {
+  return (
+    <div>
+      <h1>esta es la app para reservar</h1>
+    </div>
+  )
+}
+
+export default alquiler
